all modules done

+++ main.py
@@ -11,7 +11,7 @@ sys.path.append("deepfake-audio-generator")
 sys.path.append("deepfake-audio-generator/Real-Time-Voice-Cloning")
 sys.path.append("deepfake-video-generator")
 
-from mesoModel import predict_image
+# from mesoModel import predict_image
 from ImageSwap import detect_faces, swap_all, swap_one
 from AudioGenerator import generate_audio
 from videogenerator import deep_fake_animation
@@ -23,8 +23,8 @@ import matplotlib.pyplot as plt
 st.set_page_config(layout="wide")
 st.title("Deepfake Project")
 
-'''
-def get_prediction(image_path):
+
+'''def get_prediction(image_path):
     # Charger l'image
     open_image = cv2.imread(image_path)
     open_image = cv2.cvtColor(open_image, cv2.COLOR_BGR2RGB)
@@ -139,9 +139,18 @@ def detector_mode():
         # Affiche l'image téléchargée
         st.image("temp_image.jpg")
 
-        st.text(get_prediction("temp_image.jpg"))
+        if st.button("Predict"):
 
-        os.remove("temp_image.jpg")'''
+            progress_bar = st.progress(0)
+
+            # Mettre à jour la barre de progression pendant que le modèle génère la prédiction
+            for percent_complete in range(0, 101, 10):
+                progress_bar.progress(percent_complete)
+                time.sleep(0.5)
+
+            st.text(get_prediction("temp_image.jpg"))
+
+            os.remove("temp_image.jpg")'''
 
 
 def plot_faces(img, faces):
@@ -177,14 +186,14 @@ def plot_faces(img, faces):
 
 
 def save_img():
-    if st.button("Save Result"):
-        with open("result.jpeg", "rb") as file:
-            st.download_button(
-                label="Download Result",
-                data=file,
-                file_name="result.jpeg",
-                mime="image/jpeg",
-            )
+
+    with open("result.jpeg", "rb") as file:
+        st.download_button(
+            label="Download Result",
+            data=file,
+            file_name="result.jpeg",
+            mime="image/jpeg",
+        )
 
 
 def swap_mode():
@@ -196,6 +205,8 @@ def swap_mode():
     uploaded_image = st.file_uploader(
         "Upload your own image to test the model:", type=["jpg", "jpeg", "png"]
     )
+
+    res = None
 
     # when an image is uploaded, display image and run inference
     if uploaded_image is not None:
@@ -237,12 +248,14 @@ def swap_mode():
                     )
 
                     if selected_face_index is not None:
-                        res = swap_one(img, faces[0], faces2[selected_face_index])
+                        if st.button("Generate"):
+                            res = swap_one(img, faces[0], faces2[selected_face_index])
                     else:
                         res = None
 
                 elif len(faces2) == 1:
-                    res = swap_one(img, faces[0], faces2[0])
+                    if st.button("Generate"):
+                        res = swap_one(img, faces[0], faces2[0])
                 else:
                     res = None
 
@@ -278,14 +291,15 @@ def swap_mode():
                 if select_option == "All image":
 
                     if selected_face_index is not None:
-                        res = swap_all(img, faces, selected_face_index)
-                        cv2.imwrite("result.jpeg", res)
+                        if st.button("Generate"):
+                            res = swap_all(img, faces, selected_face_index)
+                            cv2.imwrite("result.jpeg", res)
 
-                        st.subheader("Result")
-                        st.image("result.jpeg")
+                            st.subheader("Result")
+                            st.image("result.jpeg")
 
-                        save_img()
-                        os.remove("result.jpeg")
+                            save_img()
+                            os.remove("result.jpeg")
 
                 elif select_option == "Select Another Image":
 
@@ -315,16 +329,20 @@ def swap_mode():
                             )
 
                             if selected_face_index2 is not None:
-                                res = swap_one(
-                                    img,
-                                    faces[selected_face_index],
-                                    faces2[selected_face_index2],
-                                )
+                                if st.button("Generate"):
+                                    res = swap_one(
+                                        img,
+                                        faces[selected_face_index],
+                                        faces2[selected_face_index2],
+                                    )
                             else:
                                 res = None
 
                         elif len(faces2) == 1:
-                            res = swap_one(img, faces[selected_face_index], faces2[0])
+                            if st.button("Generate"):
+                                res = swap_one(
+                                    img, faces[selected_face_index], faces2[0]
+                                )
                         else:
                             res = None
 
@@ -362,80 +380,120 @@ def voice_generator_mode():
 
         text = st.text_input("text", placeholder="Write sentence here !")
 
-        generate_audio(text, "temp_audio." + audio_file.name.split(".")[-1])
+        if text != "":
 
-        st.subheader("Result")
+            if st.button("Generate"):
 
-        st.audio("output.wav")
+                generate_audio(text, "temp_audio." + audio_file.name.split(".")[-1])
 
-        os.remove("output.wav")
-        os.remove("temp_audio." + audio_file.name.split(".")[-1])
+                st.subheader("Result")
 
+                st.audio("output.wav")
+
+                os.remove("temp_audio." + audio_file.name.split(".")[-1])
+
+                with open("output.wav", "rb") as audio_file:
+                    st.download_button(
+                        "Download Generated Audio",
+                        audio_file,
+                        file_name="output.wav",
+                        mime="audio/wav",
+                    )
+
+                os.remove("output.wav")
 
 
 def video_generation():
 
     # chemins des fichiers en local
-    config_file = 'deepfake-video-generator/config.yml'  # remplacer par le chemin réel du fichier config
-    model_weights = 'deepfake-video-generator/model_weights.tar'  # remplacer par le chemin réel du fichier config
+    config_file = "deepfake-video-generator/config.yml"  # remplacer par le chemin réel du fichier config
+    model_weights = "deepfake-video-generator/model_weights.tar"  # remplacer par le chemin réel du fichier config
 
     # interface utilisateur pour uploader les fichiers
     st.header("DeepFake Video Generator Mode")
-    st.subheader("Upload an Image and a video to Make a Deepfake")
+    st.subheader("Upload an Image to Make a Deepfake")
     source_image = st.file_uploader("Upload Source Image", type=["png", "jpg", "jpeg"])
-    driver_video = st.file_uploader("Upload Driver Video", type=["mp4"])
 
-    # specification des nomms des fichiers de sortie
-    output_video = 'generated_video.mp4'
-    output_video_fast = 'generated_video_fast.mp4'
+    if source_image is not None:
+        st.image(source_image)
 
-    # barre de progrewssion
-    progress_bar = st.progress(0)
+        st.subheader("Upload a video to Make a Deepfake")
+        driver_video = st.file_uploader("Upload Driver Video", type=["mp4"])
 
-    def generate_animation(source_path, driver_path):
-        # execution de la fonction de generation 
-        deep_fake_animation(source_path, driver_path, config_file, model_weights, output_video, output_video_fast)
-        # appelé une fois que la generation est terminée
-        progress_bar.empty()  # nettoie la barre de progression
+        if driver_video is not None:
 
-    # bouton pour lancer le processus de generation de la videao
-    if st.button('Generate Animation'):
-        if source_image and driver_video:
-            # enregistrement des fichiers uploadés sur le serveur
-            with open('temp_source_image.png', 'wb') as f:
-                f.write(source_image.getvalue())
-            st.image("temp_source_image.png")
-            with open('temp_driver_video.mp4', 'wb') as f:
-                f.write(driver_video.getvalue())
-            st.video("temp_driver_video.mp4")
+            st.video(driver_video)
 
-            # demarage du processus de generation dans un thread sseparé
-            thread = threading.Thread(target=generate_animation, args=('temp_source_image.png', 'temp_driver_video.mp4',))
-            thread.start()
+            # specification des nomms des fichiers de sortie
+            output_video = "generated_video.mp4"
+            output_video_fast = "generated_video_fast.mp4"
 
-            # tant que le thread est actif, maj de la barre de progression
-            while thread.is_alive():
-                time.sleep(0.1)  # maj toutes les 0.1 secondes
-                progress_bar.progress(50)  # maj avec la progression réelle si possible
-            thread.join()  # attendre que le trhead soit termine
+            # barre de progrewssion
+            progress_bar = st.progress(0)
 
-            # affichage de la vidéo generee si elle est disponible
-            if os.path.isfile(output_video_fast):
-                st.video(output_video_fast)
-                # utilisation de la fonction de sauvegarde pour télécharger la vidéo
-                with open(output_video_fast, 'rb') as video_file:
-                    st.download_button('Download Generated Video', video_file, file_name=output_video_fast, mime='video/mp4')
-                
-                os.remove(source_image)
-                os.remove(driver_video)
-                os.remove(output_video_fast)
-                os.remove(video_file)
-                
-        else:
-            st.error("Please upload the source image and driver video to proceed.")
+            def generate_animation(source_path, driver_path):
+                # execution de la fonction de generation
+                deep_fake_animation(
+                    source_path,
+                    driver_path,
+                    config_file,
+                    model_weights,
+                    output_video,
+                    output_video_fast,
+                )
+                # appelé une fois que la generation est terminée
+                progress_bar.empty()  # nettoie la barre de progression
 
+            # bouton pour lancer le processus de generation de la videao
+            if st.button("Generate Animation"):
+                if source_image and driver_video:
+                    # enregistrement des fichiers uploadés sur le serveur
+                    with open("temp_source_image.png", "wb") as f:
+                        f.write(source_image.getvalue())
+                    # st.image("temp_source_image.png")
+                    with open("temp_driver_video.mp4", "wb") as f:
+                        f.write(driver_video.getvalue())
+                    # st.video("temp_driver_video.mp4")
 
+                    # demarage du processus de generation dans un thread sseparé
+                    thread = threading.Thread(
+                        target=generate_animation,
+                        args=(
+                            "temp_source_image.png",
+                            "temp_driver_video.mp4",
+                        ),
+                    )
+                    thread.start()
 
+                    # tant que le thread est actif, maj de la barre de progression
+                    while thread.is_alive():
+                        time.sleep(0.1)  # maj toutes les 0.1 secondes
+                        progress_bar.progress(
+                            50
+                        )  # maj avec la progression réelle si possible
+                    thread.join()  # attendre que le trhead soit termine
+
+                    # affichage de la vidéo generee si elle est disponible
+                    if os.path.isfile(output_video_fast):
+                        st.video(output_video_fast)
+                        # utilisation de la fonction de sauvegarde pour télécharger la vidéo
+                        with open(output_video_fast, "rb") as video_file:
+                            st.download_button(
+                                "Download Generated Video",
+                                video_file,
+                                file_name=output_video_fast,
+                                mime="video/mp4",
+                            )
+
+                        os.remove("temp_source_image.png")
+                        os.remove("temp_driver_video.mp4")
+                        os.remove(output_video_fast)
+                        # os.remove(video_file)
+
+                else:
+                    st.error(
+                        "Please upload the source image and driver video to proceed."
+                    )
 
 
 page = st.sidebar.selectbox(
@@ -449,11 +507,11 @@ page = st.sidebar.selectbox(
 )
 
 if page == "DeepFake Image Detector Mode":
-    #detector_mode()
+    # detector_mode()
     pass
 elif page == "DeepFake Image Generator Mode":
     swap_mode()
 elif page == "DeepFake Audio Generator Mode":
     voice_generator_mode()
 elif page == "DeepFake Video Generator Mode":
-    pass
+    video_generation()
